Rename namespace of STL classes to AtCoder.Stl
