+++ Test/AtCoderLibrary.Test/STL/NextPermutationTest.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Xunit;
 
-namespace AtCoder
+namespace AtCoder.Stl
 {
     public class NextPermutationTest
     {
